requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# GTFS and Transit Data
gtfs-realtime-bindings>=1.0.0
protobuf>=4.23.0

# Geospatial Processing
geoalchemy2>=0.13.0
shapely>=2.0.0
fiona>=1.9.0
pyproj>=3.5.0
folium>=0.14.0

# Machine Learning
tensorflow>=2.13.0
torch>=2.0.0
xgboost>=1.7.0
scikit-learn>=1.3.0
shap>=0.42.0

# Feature Store
feast>=0.32.0

# API and Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0
streamlit-folium>=0.13.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Monitoring and Logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Cloud and Infrastructure
boto3>=1.28.0
docker>=6.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development Tools
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Simulation
simpy>=3.0.0

# Time Series
statsmodels>=0.14.0

# Configuration
pydantic>=2.0.0
python-dotenv>=1.0.0 